Shrink `LayoutMode::Grid` index from `u32` to `u16`
Adds a release-mode `assert!` guarding the 65 535-panel cap, and
documents the limit in code and `docs/grid.md`. Also adds the
`assert!`-vs-`debug_assert!` convention to `CLAUDE.md`.

+++ src/element.rs
@@ -19,8 +19,9 @@ pub enum LayoutMode {
     Canvas,
     /// WPF-style grid. Carries an index into `Tree::grid_defs` holding the row
     /// and column track definitions and per-axis gaps. Children declare cell +
-    /// span via `grid`.
-    Grid(u32),
+    /// span via `grid`. Cap is 65 535 grids per frame (`grid_defs` is cleared
+    /// each frame).
+    Grid(u16),
 }
 
 /// Per-node config: identity + spatial layout + interaction + paint flags.
